Create readme.txt
A readme file (text or markdown file) to specify the function of each component (file/folder) in the codebase,  List all dependencies used in the project, and provide instructions on installing any dependencies (e.g., pytorch, cudatoolkit, scikit-learn, etc.) If a project is built on an existing code base, it must be clearly credited and differences should be explicitly stated in the readme file. 

These details can be in a single readme file or in separate text files as convenient. Note that well-documented and clean code will be scored higher.